+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v0.19.5] - 2020-11-30
+
+### Added
+
+- FacebookSDK.tsx (unimplemented)
+- GoogleSDK.tsx (unimplemented)
+
 ## [v0.19.4] - 2020-11-30
 
 ### Changed
@@ -291,6 +298,7 @@ New:
 
 ### Added
 
+[v0.19.5]: https://github.com/thiskevinwang/coffee-code-climb/compare/v0.19.4...v0.19.5
 [v0.19.4]: https://github.com/thiskevinwang/coffee-code-climb/compare/v0.19.3...v0.19.4
 [v0.19.3]: https://github.com/thiskevinwang/coffee-code-climb/compare/v0.19.2...v0.19.3
 [v0.19.2]: https://github.com/thiskevinwang/coffee-code-climb/compare/v0.19.1...v0.19.2